Improve fit from first print
Move the standoffs one shell further in X; the previous measurements
did not account for the thickness of the shell. Make the clips on
the standoffs much smaller with gentler slope; they were too thick
and and prevented the board from fitting without cutting away.

Lots of small modifications around Ethernet and USB ports to fit
with more variations. The reduction in pcb board thickness is
part of this.

Change all fonts to bold as the horizontal parts mostly disappeared
in the layers.  Adjust some off the offsets to more clearly label
the stepper motor connectors.

Two of the top/bottom case clips broke on initial assembly, so make
the engagement between the case halves much smaller.