Improve fit from second print
The bottom of the molex connector is lower than the top of the
ethernet jack, so increase the size of the lower case half and
cut away a slot for the molex connector.  The mini/micro USB
slot did not reach quite high enough to be useful, so move it
up to make it more likely to work.

Change the clip structure to a tab and relief instead of interlocking
tabs, and make the clip thicker and stronger.  Clips still broke
on the second iteration.

The example gcode file is sliced with .15mm layers as this seems
to produce better results for labels on top half.  Switched from
3 to 5 top and bottom solid layers to accomodate this change.