Improve fit from third print
Center the board properly instead of trying to use measured offsets.

Accomodate the BB ethernet and Replicape inductor height with a
cutout from the case top instead of increasing the bottom height,
because increasing the bottom height interfered with all the side
connectors.

Reduce board standoff clip size further; 1mm worked for PETG but
is too much for PLA+, so try 0.5mm.

Add some slop for interference fit for top clips; make the relief
larger and cut away the edge of the tab to make it easier to insert
the clips into the case.  Use mirror() to distribute the clips.
(Should probably use mirror() a lot more in the whole design!)

Change the power labels; the small font did not print suffiently
legibly even in bold with 0.15mm layers, so use a larger font print
only one character vertically