Fix off-center mounting from fourth print
The previous print inadvertently moved the BB out of center instead
of closer to center. Fix that.

Change the labels for the power connectors to be oriented vertically
like the other labels, and simplify to use one character for power
connections that are not redundant.

Change the fan hole to use more segments to make it smoother.

Update the README with instructions for installing the BB.