Fix label mistake from fifth print
The heated bed and second extruder heater labels had the wrong
number of letters, so fixed them to follow the pattern that the
doubled connectors have doubled letters.

Made the clips even thicker to make them stronger, since I broke
one clip again when assembling the fifth print.  Made them as
thick as they can reasonably be in the Y axis by matching the
cutout for the molex connector.  To strengthen them more would
require a change in design.

No change to the bottom box, only the top.